pyproject.toml
[project]
name = "anamorpher"
version = "0.1.0"
description = "Image scaling and adversarial generation tool"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "Flask==3.0.0",
    "Flask-CORS==4.0.0",
    "opencv-python==4.8.1.78",
    "torch==2.1.0",
    "torchvision==0.16.0",
    "tensorflow==2.14.0",
    "Pillow==10.1.0",
    "numpy==1.24.3",
    "bleach==6.1.0",
    "MarkupSafe==2.1.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "toml>=0.10.2",
    "ruff>=0.1.0",
]